requirements.txt
# GM Services Requirements
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7  # PostgreSQL adapter
pymysql==1.1.0          # MySQL adapter

# Forms and validation
WTForms==3.1.0
email-validator==2.1.0

# Authentication and security
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
cryptography==41.0.7

# Payment processing
stripe==7.8.0
paypalrestsdk==1.13.3

# Real-time features
python-socketio==5.10.0
eventlet==0.33.3

# HTTP requests
requests==2.31.0
urllib3==2.1.0

# Date and time
python-dateutil==2.8.2

# Image processing
Pillow==10.1.0

# Utilities
python-slugify==8.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.7

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
flake8==6.1.0
black==23.11.0

# Production deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Environment and configuration
python-decouple==3.8

# OAuth
Authlib==1.2.1
requests-oauthlib==1.3.1

# Rate limiting
Flask-Limiter==3.5.0

# Admin interface (optional)
Flask-Admin==1.6.1

# Caching
Flask-Caching==2.1.0

# Monitoring and logging
sentry-sdk[flask]==1.38.0

# API documentation
flasgger==0.9.7.1

# Data processing and file handling
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1